+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### Added
 
+* [HexaPDF::Layout::LineFragment] for single text line layout calculations
 * [HexaPDF::Layout::TextShaper] for text shaping functionality
 * [HexaPDF::Layout::TextFragment] for basic text metrics calculations
+* [HexaPDF::Layout::InlineBox] for fixed size inline graphics
 * Support for kerning of TrueType fonts via the 'kern' table
 * Support for determining the features provided by a font